Refactor in order to improve readbility in analysis queries
